refactor: move isBrowser into utils folder
