+++ test/playground/pages/dynamic-import-file-env/hello.server.ts
@@ -1,3 +1,3 @@
-import { isBrowser } from '../isBrowser'
+import { isBrowser } from '../../utils/isBrowser'
 if (isBrowser) throw new Error("I shouldn't be imported from the client")
 console.log('hello from server')
