--- test/playground/pages/isBrowser.ts
@@ -1,1 +1,0 @@
-export const isBrowser = typeof window === 'object'
